feat: agent tool to query card db using the API and moved Pydantic models to models.py
